+++ config/auth-log.php
@@ -15,10 +15,10 @@ return [
         //add credentials here to be ommitted from being stored in the log database
     ],
     'credentialsToEncrypt' => [
-        'password'
+        // 'password'
     ],
     'fieldsToOmit' => [
-        'user_ip',
+        // 'user_ip',
         //add fields here to be omitted from being stored in the log database
     ]
 ];